+++ apps/mobile/src/components/layouts/tabbar/BottomTabs.tsx
@@ -1,86 +1,25 @@
-import type { BottomTabBarProps } from "@react-navigation/bottom-tabs"
-import type { ForwardRefExoticComponent } from "react"
-import { forwardRef, useMemo, useRef, useState } from "react"
-import type { ScrollView } from "react-native"
-import { useSharedValue } from "react-native-reanimated"
+import { useContext, useRef } from "react"
 
-import { BottomTabHeightProvider } from "./BottomTabHeightProvider"
-import {
-  AttachNavigationScrollViewContext,
-  SetAttachNavigationScrollViewContext,
-} from "./contexts/AttachNavigationScrollViewContext"
 import { BottomTabBarBackgroundContext } from "./contexts/BottomTabBarBackgroundContext"
-import {
-  BottomTabBarVisibleContext,
-  SetBottomTabBarVisibleContext,
-} from "./contexts/BottomTabBarVisibleContext"
 import { useNavigationScrollToTop } from "./hooks"
 import { Tabbar } from "./Tabbar"
-// import { Tabs } from "./TabClient"
-
-type ExtractReactForwardRefExoticComponent<T> =
-  T extends React.ForwardRefExoticComponent<infer P> ? P : never
 
 export const BottomTabs = () => {
-  const opacity = useSharedValue(1)
-  const [tabBarVisible, setTabBarVisible] = useState(true)
-  const [attachNavigationScrollViewRef, setAttachNavigationScrollViewRef] =
-    useState<React.RefObject<ScrollView> | null>(null)
-
   const currentIndex = useRef<number | undefined>(undefined)
-  const scrollToTop = useNavigationScrollToTop(attachNavigationScrollViewRef)
+  const scrollToTop = useNavigationScrollToTop()
+  const { opacity } = useContext(BottomTabBarBackgroundContext)
   return (
-    <AttachNavigationScrollViewContext.Provider value={attachNavigationScrollViewRef}>
-      <SetAttachNavigationScrollViewContext.Provider value={setAttachNavigationScrollViewRef}>
-        <BottomTabBarBackgroundContext.Provider value={useMemo(() => ({ opacity }), [opacity])}>
-          <SetBottomTabBarVisibleContext.Provider value={setTabBarVisible}>
-            <BottomTabBarVisibleContext.Provider value={tabBarVisible}>
-              <BottomTabHeightProvider>
-                {/* <Tabs
-                  {...props}
-                  tabBar={TabBar}
-                  screenListeners={{
-                    ...props.screenListeners,
-                    tabPress: (e) => {
-                      if (props.screenListeners && "tabPress" in props.screenListeners) {
-                        props.screenListeners.tabPress!(e)
-                      }
+    <Tabbar
+      onPress={(index) => {
+        opacity.value = 1
 
-                      if (currentTarget.current === e.target) {
-                        scrollToTop()
-                        return
-                      }
+        if (currentIndex.current === index) {
+          scrollToTop()
+          return
+        }
 
-                      opacity.value = 1
-                      currentTarget.current = e.target
-                    },
-                    transitionStart: (e) => {
-                      if (props.screenListeners && "transitionStart" in props.screenListeners) {
-                        props.screenListeners.transitionStart!(e)
-                      }
-                      currentTarget.current = e.target
-                      opacity.value = 1
-                    },
-                  }}
-                  ref={ref}
-                /> */}
-                <Tabbar
-                  onPress={(index) => {
-                    opacity.value = 1
-
-                    if (currentIndex.current === index) {
-                      scrollToTop()
-                      return
-                    }
-
-                    currentIndex.current = index
-                  }}
-                />
-              </BottomTabHeightProvider>
-            </BottomTabBarVisibleContext.Provider>
-          </SetBottomTabBarVisibleContext.Provider>
-        </BottomTabBarBackgroundContext.Provider>
-      </SetAttachNavigationScrollViewContext.Provider>
-    </AttachNavigationScrollViewContext.Provider>
+        currentIndex.current = index
+      }}
+    />
   )
 }
